backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
distro==1.9.0
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.24.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.13
numpy==2.3.1
openai==1.93.0
packaging==25.0
pillow==11.3.0
preshed==3.0.10
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dotenv==1.1.1
requests==2.32.4
rich==14.0.0
rsa==4.9.1
setuptools==80.9.0
shellingham==1.5.4
smart_open==7.3.0
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
tenacity==8.5.0
thinc==8.3.6
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.2
